+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### March
 
+* March 31 - Admin dashboard for referrals #799 @hafezbusra
 * March 30 - Affiliate program landing page and sign up form #850
 * March 28 - Refactor `Hired::Form` into `Developers::CelebrationPackageRequest` #836 @MikeRayUX
 * March 25 - Add developer scheduling link #838 @jkostolansky